Refactor: PR 리뷰 반영 및 리팩토링 진행
