+++ frontend/sidedish/src/components/Header.jsx
@@ -1,11 +1,25 @@
 import styled from "styled-components";
 import Gnb from "./Gnb";
 
+const Header = () => {
+  return (
+    <HeaderContainer>
+      <Home>Ordering</Home>
+      <Gnb />
+      <Blank />
+      <Img src="./images/searchIcon.png" alt="돋보기아이콘" />
+      <Img src="./images/userIcon.png" alt="사람아이콘" />
+      <Img src="./images/cartIcon.png" alt="카트아이콘" />
+    </HeaderContainer>
+  );
+};
+
 const HeaderContainer = styled.header`
   display: flex;
   flex-direction: row;
   padding: 16px 80px;
   border-bottom: 1px solid black;
+  white-space: nowrap;
 `;
 
 const Home = styled.h1`
@@ -27,18 +41,5 @@ const Img = styled.img`
   padding: 20px 13px;
   cursor: pointer;
 `;
-
-const Header = () => {
-  return (
-    <HeaderContainer>
-      <Home>Ordering</Home>
-      <Gnb />
-      <Blank />
-      <Img src="./images/searchIcon.png" alt="돋보기아이콘" />
-      <Img src="./images/userIcon.png" alt="사람아이콘" />
-      <Img src="./images/cartIcon.png" alt="카트아이콘" />
-    </HeaderContainer>
-  );
-};
 
 export default Header;
