+++ frontend/sidedish/src/components/ItemCards.jsx
@@ -1,11 +1,6 @@
 import styled from "styled-components";
 import ItemCard from "./ItemCard";
 
-const ItemCardContainer = styled.ul`
-  display: flex;
-  padding: 34px 56px;
-`;
-
 const ItemCards = ({ dataState, len }) => {
   const itemCardList = dataState.map(({ image, title, description, n_price, s_price, badge }, ind) => (
     <ItemCard key={ind} image={image} title={title} description={description} n_price={n_price} s_price={s_price} badge={badge} len={len}></ItemCard>
@@ -13,5 +8,10 @@ const ItemCards = ({ dataState, len }) => {
 
   return <ItemCardContainer>{itemCardList}</ItemCardContainer>;
 };
+
+const ItemCardContainer = styled.ul`
+  display: flex;
+  padding: 34px 56px;
+`;
 
 export default ItemCards;
